Separate value object tests
